# config/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "v2",
        "title": "REST API V1"
    },
    "host": "localhost:8080",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/2/subscriptions/{DEVICE_ID}": {
            "put": {
                "tags": [
                    "subscriptions"
                ],
                "description": "Update subscription object, must include at least 1 parameter to update. (last session date and sessions count wil be updated with this method).",
                "operationId": "updateSub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "x-pushbots-appid",
                        "type": "string",
                        "description": "applicaiton Id for which the device belongs, can be found in application settings in dashboard.",
                        "required": true
                    },
                    {
                        "in": "header",
                        "name": "Content-Type",
                        "type": "string",
                        "description": "application/json",
                        "required": true
                    },
                    {
                        "name": "DEVICE_ID",
                        "in": "path",
                        "type": "string",
                        "description": "Valid Object Id, can be replaced with token ,The device Id or token to update, it must include at least 1 parameter to update. The application should store the object Id, in order to be able to update device data or sessions. token can be used to update device data, however it's always recommended to use device Id instead.",
                        "required": true
                    },
                    {
                        "in": "body",
                        "name": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateSubscription"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated: {\"_id\": \"56d0a0deaa760e1711a44e18,\"sessions\": \"26\",\"last_session\": \"2016-02-26T19:01:48.662Z\"}"
                    }
                }
            },
            "get": {
                "tags": [
                    "subscriptions"
                ],
                "description": "Retrieve Subscription.",
                "operationId": "getSub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "x-pushbots-appid",
                        "type": "string",
                        "description": "applicaiton Id for which the device belongs, can be found in application settings in dashboard."
                    },
                    {
                        "name": "DEVICE_ID",
                        "in": "path",
                        "type": "string",
                        "description": "Valid Object Id, can be replaced with token ,The device Id or token to update, it must include at least 1 parameter to update. The application should store the object Id, in order to be able to update device data or sessions. token can be used to update device data, however it's always recommended to use device Id instead.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated: {\"appID\" : \"583be4ef357e71cd090041a8\", \"token\" : \"ca8d88f2e3299cedd8c770fc051b8b19a18bf62b8fdd1c202d6d738bef069094\", \"sessions\" : 38, \"m\" : { \"jb\" : false, \"lib\" : \"2.0.4\", \"l\" : \"de_EG\", \"appv\" : \"1.0\", \"r\" : \"640x1136\", \"d\" : \"iPhone6,2\", \"osv\" : \"10.1.1\" }, \"ccode\" : \"EG\", \"pp\" : 2, \"tz\" : \"7200\", \"ls\" : ISODate(\"2016-12-25T19:08:35.919Z\"), \"platform\" : \"0\",\" __v\" : 0, \"badg\" : 72, \"alias\" : \"test\" }"
                    },
                    "401": {
                        "description": "Failed: \"Device not found\""
                    }
                }
            },
            "delete": {
                "tags": [
                    "subscriptions"
                ],
                "description": "Delete Subscription from PushBots.",
                "operationId": "deleteSub",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "x-pushbots-appid",
                        "type": "string",
                        "description": "applicaiton Id for which the device belongs, can be found in application settings in dashboard.",
                        "required": true
                    },
                    {
                        "name": "DEVICE_ID",
                        "in": "path",
                        "type": "string",
                        "description": "Valid Object Id, can be replaced with token ,The device Id or token to update, it must include at least 1 parameter to update. The application should store the object Id, in order to be able to update device data or sessions. token can be used to update device data, however it's always recommended to use device Id instead.",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success:  {\"deleted\": true, \"id\": \"56d0a0deaa760e1711a44e18\"}"
                    },
                    "401": {
                        "description": "Failed: {\"deleted\":false}"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "UpdateSubscription": {
            "type": "object",
            "required": [
                "uuid",
                "token",
                "platform"
            ],
            "properties": {
                "uuid": {
                    "type": "string",
                    "description": "Unique device Id."
                },
                "token": {
                    "type": "string",
                    "description": "The Apple or Google generated token used to deliver messages to the APNs or GCM push networks respectively."
                },
                "platform": {
                    "type": "integer",
                    "description": "The type of device, 0 for iOS devices, 1 for Android devices, 2 for Chrome users, 3 for firefox users."
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "description": "An array of the tags to which a device is currently subscribed."
                },
                "tags_add": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "tags_remove": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "timeZone": {
                    "type": "string",
                    "description": "The offset of the time zone from UTC, ignoring daylight savings time e.g. 7200."
                },
                "alias": {
                    "type": "string",
                    "description": "device alias or username."
                },
                "countryCode": {
                    "type": "integer",
                    "description": "Country code to detect country generated by carrier."
                },
                "badge": {
                    "type": "integer",
                    "description": "is a number field representing the last known application badge for iOS installations only."
                },
                "debug": {
                    "type": "boolean",
                    "description": "Set debug mode for the device."
                },
                "badge_p": {
                    "type": "integer",
                    "description": "is a number field to be incremented to the badge for iOS installations only."
                },
                "badge_d": {
                    "type": "integer",
                    "description": "is a number field to be substracted from the badge for iOS installations only."
                },
                "location": {
                    "type": "array",
                    "description": "device location [lat, lng].",
                    "example": [
                        50,
                        30
                    ],
                    "items": {
                        "type": "integer"
                    }
                },
                "metrics": {
                    "$ref": "#/definitions/DeviceData",
                    "description": "Object with device data for better analytics tracking. e.g. {\"appVersion\":\"1.1\",\"resolution\":\"396x240\",\"deviceModel\":\"Android 5.0.0 - API 21 - 768x1280\",\"locale\":\"en_US\",\"PushBotslibVersion\":\"3.0.1\",\"osVersion\":\"5.0\",\"carrier\":\"Vodafone\"}"
                }
            }
        },
        "DeviceData": {
            "type": "object",
            "properties": {
                "carrier": {
                    "type": "string",
                    "example": "Vodafone",
                    "description": "device carrier e.g. Vodafone."
                },
                "osVersion": {
                    "type": "string",
                    "description": "9.1.2 for iOS/ 4.4.4 for Android.",
                    "example": "9.1.2"
                },
                "resolution": {
                    "type": "string",
                    "example": "396x240"
                },
                "appVersion": {
                    "type": "string",
                    "description": "your application versioning for iOS AppStore or Google playstore."
                },
                "locale": {
                    "type": "string",
                    "example": "en_US",
                    "description": "device language"
                },
                "PushBotslibVersion": {
                    "type": "string",
                    "description": "Pushbots library version used in library"
                },
                "jailbroken": {
                    "type": "boolean",
                    "description": "If the device is jailbroken or no"
                }
            }
        },
...
    }
}